getservice_details/requirements.txt
flask==2.3.2
google-cloud-bigquery==3.12.0
